requirements.txt
django==4.2.7
pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap4
reportlab==3.6.12
django-widget-tweaks==1.4.13

# Added for Streamlit frontend + API
djangorestframework>=3.13
django-cors-headers>=4.0
streamlit>=1.20
requests>=2.28
# Production / deployment runtime
gunicorn>=20.1
whitenoise>=6.0
supervisor>=4.2
